Updated trip_id description in rider_info.txt via MBTA
Via MBTA "Trip_id is identified as required, but there are cases in which trip_id cannot be determined and information for all other fields is available, so making this field optional lets agencies publish more data. An ideal example would be WMATA Metro service, on which customers with SmarTrip cards tap to enter the system and tap to exit. Which trip was taken, or which lines were taken and how many transfers there were, are unknown. 

The following language is proposed: 

Optional. The trip_id field contains the ID of the vehicle trip associated with the unique rider, if this is known. If trip_id is empty then the rider may have taken one of several trips, or several possible chains of trips, to travel between the origin and the destination."