added record_use field
Added record_use field to highlight the fact that either boarding counts, alighting counts, or schedule cancellation information should be including. This means that entries with no ridership counts should be considered valid only there is a service cancellation indicated in schedule_relationship. This is also attempting to address the issue of requiring some information to be entered in the optional ridership count fields (except in the scenario described above).