Moved schedule_relationship
Reordered schedule_relationship to come after the new record_use field and before boardings and alightings. This is for the purpose of qualifying the following ridership data, and in attempts to highlight its importance as one of the three "expected" (but not explicitly required) fields, along with boardings and alightings.