added current_count to board_alight.txt
load_count allows the specification of the count of riders onboard a vehicle at a given stop. Its inclusion in addition to current_load is needed as capacity may not be available for a vehicle and the start of a trip may not have zero load. If the start of the trip does not have zero load and capacity is not known, rider counts onboard a vehicle are not possible without this field.